Move data structure into sql_models
